# Parallelize Parquet file reading on-node