@bmcdonald3 2022-01-27
On doing some more testing, I discovered an off by one error when reading more than 20 files, will need to explore that more before merging.