@bmcdonald3 2022-01-27
This was actually an issue that I had already encountered in the Chapel code; in the unit test I was appending the number straight to the end of the file name, which caused "file10" to come before "file1", causing the read ins to come back in a weird order, so it was a bug with the test, not the parallelization code.