@reuster986 2022-02-01
@ronawho @bmcdonald3 This is awesome! And yes I did miss it -- I have like 4 different jobs these days, so thank you for pinging again!

@mhmerrill @glitch once #1055 gets merged (since it's a correctness fix), I think we should cut another release to test parquet read performance internally.